[JAVA-16207] Moved Json related articles from kotlin-libraries-2 to new kotlin-json module
