+++ kotlin-json/src/main/kotlin/com/baeldung/kotlin/gson/Article.kt
