+++ kotlin-json/src/main/kotlin/com/baeldung/kotlin/gson/Author.kt
