+++ kotlin-json/src/main/kotlin/com/baeldung/kotlin/jackson/Book.kt
