+++ kotlin-json/src/test/resources/com/baeldung/kotlin/moshi/list_of_employees.json
