+++ kotlin-json/src/test/resources/com/baeldung/kotlin/moshi/single_salary_record.json
