+++ kotlin-json/src/test/resources/com/baeldung/kotlin/moshi/snake_project.json
