+++ kotlin-libraries-2/README.md
@@ -5,16 +5,11 @@ This module contains articles about Kotlin Libraries.
 ### Relevant articles:
 
 - [gRPC and Kotlin](https://www.baeldung.com/kotlin/grpc)
-- [Jackson Support for Kotlin](https://www.baeldung.com/kotlin/jackson-kotlin)
 - [Introduction to RxKotlin](https://www.baeldung.com/kotlin/rxkotlin)
 - [Kotlin Immutable Collections](https://www.baeldung.com/kotlin/immutable-collections)
 - [Dependency Injection for Kotlin with Injekt](https://www.baeldung.com/kotlin/dependency-injection-with-injekt)
 - [Introduction to Kovenant Library for Kotlin](https://www.baeldung.com/kotlin/kovenant)
-- [Converting Kotlin Data Class from JSON using GSON](https://www.baeldung.com/kotlin/json-convert-data-class)
-- [Iterate Through a JSONArray in Kotlin](https://www.baeldung.com/kotlin/iterate-over-jsonarray)
-- [Parsing JSON Arrays in Kotlin with Gson](https://www.baeldung.com/kotlin/gson-parse-arrays)
 - [Convert a Kotlin data class object to a Map](https://www.baeldung.com/kotlin/data-class-to-map)
 - [Read and Write CSV Files With Kotlin](https://www.baeldung.com/kotlin/csv-files)
 - [Kotlin Coroutines and RxKotlin Comparison](https://www.baeldung.com/kotlin/coroutines-vs-rxkotlin)
-- [Moshi JSON Library for Kotlin Applications](https://www.baeldung.com/kotlin/moshi-json-library)
 - More articles: [[<-- prev]](/kotlin-libraries)
